OC-206: resolve invalid merge
