+++ clover-core/src/test/groovy/com/atlassian/clover/JavaSyntax15CompilationTest.groovy
@@ -6,52 +6,30 @@ import org.junit.Test
 
 /**
  * The purpose of this test is to
- * <li>make sure the code compiles under JDK1.5 or later</li>
- * <li>make sure that when that code is instrumented, it still compiles</li>
+ * a) make sure the code compiles under a JDK15
+ * b) make sure that when that code is instrumented, it still compiles
  */
 class JavaSyntax15CompilationTest extends JavaSyntaxCompilationTestBase {
 
+    protected File srcDir
+
+    /** Regular expression for: __CLR_hash_code.R.inc(index) */
+    protected final String R_INC = "__CLR[a-zA-Z0-9_]+\\.R\\.inc\\([0-9]+\\);"
+
     @Before
-    void setUp() {
+    void setUp() throws Exception {
         setUpProject()
+        srcDir = new File(mTestcasesSrcDir, "javasyntax1.15")
+        resetAntOutput()
     }
 
     @Test
-    void testCompilation_15() throws Exception {
-        final File srcDir = new File(mTestcasesSrcDir, "javasyntax1.5")
-        compileSources(srcDir, JavaEnvUtils.JAVA_1_7)
+    void testTextBlock() {
+        final String fileName = "java15/Java15TextBlock.java"
+        File srcFile = new File(srcDir, fileName)
+
+        // Currently, OpenClover cannot be built on JDK15
+        instrumentSourceFile(srcFile, JavaEnvUtils.JAVA_15)
+        assertFileMatches(fileName, R_INC + "System.out.println", false)
     }
-
-    /**
-     * Test java 1.5 language features and clover handles them.
-     *
-     * @throws Exception
-     */
-    @Test
-    void testInstrumentationAndCompilation_15() throws Exception {
-        final File srcDir = new File(mTestcasesSrcDir, "javasyntax1.5")
-        instrumentAndCompileSources(srcDir, JavaEnvUtils.JAVA_1_7)
-
-        String[] testCaseMainClasses = [
-                "coverage.enums.EnumTests",
-                "coverage.metadata.TestCases",
-        ]
-        executeMainClasses(testCaseMainClasses)
-
-        // assert metrics
-        assertMethodCoverage("coverage.enums.E1", 13)
-
-        assertStatementCoverage("coverage.enums.E2", 8, 1)
-        assertMethodCoverage("coverage.enums.E2", 14, 2)
-
-        assertMethodCoverage("coverage.enums.E3", 7)
-
-        //TODO remove this once 1.5 goes final?
-//        assertMethodCoverage("coverage.metadata.Annot2", 8)
-//        assertMethodCoverage("coverage.metadata.Annot2", 16)
-
-        assertMethodCoverage("coverage.metadata.DeprecatedTest", 5)
-    }
-
 }
-
