+++ clover-core/src/test/groovy/com/atlassian/clover/JavaSyntax5CompilationTest.groovy
@@ -1,6 +1,8 @@
 package com.atlassian.clover
 
-import org.apache.tools.ant.util.JavaEnvUtils
+import com.atlassian.clover.util.JavaEnvUtils
+import org.junit.Before
+import org.junit.Test
 
 /**
  * The purpose of this test is to
@@ -9,9 +11,15 @@ import org.apache.tools.ant.util.JavaEnvUtils
  */
 class JavaSyntax5CompilationTest extends JavaSyntaxCompilationTestBase {
 
+    @Before
+    void setUp() {
+        setUpProject()
+    }
+
+    @Test
     void testCompilation_15() throws Exception {
         final File srcDir = new File(mTestcasesSrcDir, "javasyntax1.5")
-        compileSources(srcDir, JavaEnvUtils.JAVA_1_7)
+        compileSources(srcDir, JavaEnvUtils.JAVA_7)
     }
 
     /**
@@ -19,9 +27,10 @@ class JavaSyntax5CompilationTest extends JavaSyntaxCompilationTestBase {
      *
      * @throws Exception
      */
+    @Test
     void testInstrumentationAndCompilation_15() throws Exception {
         final File srcDir = new File(mTestcasesSrcDir, "javasyntax1.5")
-        instrumentAndCompileSources(srcDir, JavaEnvUtils.JAVA_1_7)
+        instrumentAndCompileSources(srcDir, JavaEnvUtils.JAVA_7)
 
         String[] testCaseMainClasses = [
                 "coverage.enums.EnumTests",
